Collapse neo4j_process into neo4j
