+++ templates/compare.html
@@ -145,7 +145,7 @@
         };
     }
     keys.forEach(create_toggle_link);
-    show_table_of(keys[0]);
+    show_table_of('all');
     function create_toggle_link(key){
         let link = document.createElement("a");
         let li = document.createElement("li");
